refactor: move ImageSlider to components dir. show entire scaled down image on mobile screens
